feat: integrate RenderEngine from @expressots/shared and remove deprecated types
