+++ package.json
@@ -46,9 +46,6 @@
     "framework",
     "server-side"
   ],
-  "engines": {
-    "node": ">=18.18.0"
-  },
   "scripts": {
     "prepare": "husky",
     "clean": "node scripts/rm.js lib",
@@ -74,7 +71,7 @@
     "@commitlint/cli": "19.2.1",
     "@commitlint/config-conventional": "19.2.2",
     "@expressots/core": "file:../core/expressots/expressots-core-2.16.2.tgz",
-    "@expressots/shared": "0.1.0",
+    "@expressots/shared": "file:../shared/expressots-shared-0.1.0.tgz",
     "@release-it/conventional-changelog": "8.0.1",
     "@types/express": "4.17.21",
     "@types/node": "20.14.10",
